+++ bindings_wasm/src/content_types/remote_attachment.rs
@@ -16,7 +16,7 @@ pub struct RemoteAttachment {
   pub nonce: Vec<u8>,
   pub scheme: String,
   #[wasm_bindgen(js_name = "contentLength")]
-  pub content_length: i64,
+  pub content_length: u32,
   pub filename: Option<String>,
 }
 
@@ -29,7 +29,7 @@ impl From<xmtp_content_types::remote_attachment::RemoteAttachment> for RemoteAtt
       salt: remote.salt,
       nonce: remote.nonce,
       scheme: remote.scheme,
-      content_length: remote.content_length as i64,
+      content_length: remote.content_length as u32,
       filename: remote.filename,
     }
   }
